making the css/js files load dynamically just by adding them in config/templateconfig .... also adding multi language feature

+++ app/views/employee/default.view.php
@@ -39,7 +39,7 @@ table tr:nth-child(odd){
                     echo "<td>" . $employee->salary."</td>";
                     echo "<td> 
                                 <a href='/mvc/employee/edit/"   . $employee->id . "'>Edit</a>
-                                <a href='/mvc/employee/delete/" . $employee->id . "'>Delete</a>";
+                                <a class='delete-confirm' href='/mvc/employee/delete/" . $employee->id . "'>Delete</a>";
                     echo "</td>";                                                           
                 echo "</tr>";     
                 }
